be more precise about data dependencies coming into nodes at which the cache state is indetermined.
we can only do this if the accessed cache line does not depend on the data value.
i.e.: we can not do this for array accesses, unless the array index is constant.

in order to be more precise, but still differentiate these two cases, we
need one additional join node in the cache cost graph.
in the (non-constant) array case, the data access will hapen at the split node (at which multiple exeucution times
are possible).
In the array case, data access happens later, after the two different timings have joined, again.